make-release-package.js: Remove destination directory only if it exists

+++ bake/make-release-package.js
@@ -36,7 +36,9 @@ function renameMidiFiles(targetDir) {
 module.exports = v => {
     const dest = 'DestinationUnderworldRelease/'
 
-    fs.rmSync(dest, { recursive: true })
+    if (fs.existsSync(dest)) {
+        fs.rmSync(dest, { recursive: true })
+    }
 
     fs.mkdirSync(`${dest}`)
     fs.mkdirSync(`${dest}dataloss`)
